requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bidict==0.22.1
bitarray==2.8.1
blinker==1.6.2
Bootstrap-Flask==2.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
cytoolz==0.12.2
distlib==0.3.6
dominate==2.8.0
eclipse-zenoh==0.10.0rc0
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
filelock==3.12.2
Flask==2.3.3
flask-bootstrap5==0.1.dev1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-WTF==1.1.1
fonttools==4.43.1
frozenlist==1.4.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kaleido==0.2.1
kiwisolver==1.4.5
lru-dict==1.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
maturin==1.3.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.9.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parsimonious==0.9.0
patsy==0.5.3
Pillow==10.1.0
platformdirs==3.5.3
plotly==5.17.0
plotly-express==0.4.1
protobuf==4.24.2
py-solc-x==1.1.1
pycryptodome==3.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
pyunormalize==15.0.0
pywin32==306
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.2
scipy==1.11.3
seaborn==0.13.0
semantic-version==2.10.0
simple-websocket==0.10.1
six==1.16.0
statsmodels==0.14.0
tenacity==8.2.3
toolz==0.12.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
varint==1.0.2
virtualenv==20.23.0
visitor==0.1.3
web3==6.9.0
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2